requirements/common.txt
djangorestframework
firebase-admin
djangorestframework-simplejwt
python-dotenv
psycopg2
dj-database-url
drf-yasg